chore: rename test directory to __tests__
