+++ src/TusClientProvider/TusClientProvider.tsx
@@ -3,14 +3,14 @@ import { useReducer } from 'react';
 import {
   TusClientDispatchContext,
   TusClientStateContext,
-} from './core/tusContexts';
+} from '../core/contexts';
 
 import {
   tusClientInitialState,
   tusClientReducer,
-} from './core/tusClientReducer';
-import { TusConfigs, TusHandler } from './core/tusHandler';
-import { TusController } from './core/TusController';
+} from '../core/tusClientReducer';
+import { TusConfigs, TusHandler } from '../core/tusHandler';
+import { TusController } from './TusController';
 
 export type TusClientProviderProps = Readonly<TusConfigs>;
