+++ src/TusClientProvider/TusController.tsx
@@ -1,8 +1,8 @@
 import { FC, useEffect } from 'react';
-import { ERROR_MESSAGES } from './constants';
-import { updateTusHandlerOptions } from './tucClientActions';
-import { useTusClientDispatch, useTusClientState } from './tusContexts';
-import { TusConfigs } from './tusHandler';
+import { ERROR_MESSAGES } from '../core/constants';
+import { updateTusHandlerOptions } from '../core/tucClientActions';
+import { useTusClientDispatch, useTusClientState } from '../core/contexts';
+import { TusConfigs } from '../core/tusHandler';
 
 export type TusControllerProps = Readonly<TusConfigs>;
