+++ src/TusClientProvider/index.ts
@@ -1,0 +1,2 @@
+export { TusClientProvider } from './TusClientProvider';
+export type { TusClientProviderProps } from './TusClientProvider';
