+++ src/useTus/index.ts
@@ -1,0 +1,2 @@
+export { useTus } from './useTus';
+export type { UseTusResult, UseTusOptions } from './useTus';
